[Core] Moving lazy_require to separated file and move which to utils
